fix(intel): update individual return result for hps and fpga bridges
The code is designed to execute SOC2FPGA and LWSOC2FPGA first then to F2SOC
and both sharing the same result "return". Thus when F2SOC is executed,
the "return" result will overwrite SOC2FPGA "return" result even though
it is not enabled. Using 2 different "return" result to for each bridges and
return both of them at the end of the function to avoid being overwritten.

Signed-off-by: Jit Loon Lim <jit.loon.lim@intel.com>